+++ install.py
@@ -9,7 +9,7 @@ import secrets
 
 from weecfg.extension import ExtensionInstaller
 
-VERSION = '0.16.0'
+VERSION = '0.17.0'
 
 
 def loader():